This reverts commit e2e3b333ce012ba9b553ec4b55dcd83b24d221db.
# Conflicts:
#	app/src/main/java/com/example/mreznikviz/animations/BounceAnimation.kt

+++ app/src/main/java/com/example/mreznikviz/WaitingFriendsActivity.kt
@@ -86,7 +86,7 @@ class WaitingFriendsActivity : AppCompatActivity() {
             FirebaseDatabase.getInstance().reference.child("quiz/" + quiz.id + "/start").setValue(viewAdapter.itemCount + 1)
             startActivity(Intent(this, QuestionActivity::class.java)
                 .putExtra("questions", quiz)
-                .putExtra("nop", viewAdapter.itemCount + 1))
+                .putExtra("nop", (viewAdapter.itemCount + 1).toLong()))
         }.enableOnTouchDemand()
 
     }
